Abstract derive route into handler & add credentials/derive.
