+++ lib/routes/index.js
@@ -1,32 +1,13 @@
 /*!
- * Copyright (c) 2023 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2023 - 2024 Digital Bazaar, Inc.
  */
-import * as vc from '@digitalbazaar/vc';
-import assert from 'assert-plus';
-import {
-  createDiscloseCryptosuite as createEcdsaSd2023DiscloseCryptosuite
-} from '@digitalbazaar/ecdsa-sd-2023-cryptosuite';
-import {DataIntegrityProof} from '@digitalbazaar/data-integrity';
+import {deriveHandler} from './handler.js';
 import express from 'express';
-import {loader} from '../documentLoader.js';
-
-const documentLoader = loader.build();
 
 export const router = express.Router();
 
-router.post('/derive', asyncHandler(async (req, res) => {
-  const {options: {selectivePointers}, verifiableCredential} = req.body;
-  assert.optionalArrayOfString(selectivePointers, 'selectivePointers');
-  assert.object(verifiableCredential, 'verifiableCredential');
-  const cryptosuite = createEcdsaSd2023DiscloseCryptosuite({selectivePointers});
-  const suite = new DataIntegrityProof({cryptosuite});
-  const derivedVC = await vc.derive({
-    verifiableCredential,
-    suite,
-    documentLoader
-  });
-  res.send({verifiableCredential: derivedVC});
-}));
+router.post('/derive', asyncHandler(deriveHandler));
+router.post('/credentials/derive', asyncHandler(deriveHandler));
 
 function asyncHandler(middleware) {
   return function asyncMiddleware(...args) {
